rm paddlelite in setup.py
