+++ setup.py
@@ -75,7 +75,6 @@ base = [
     "braceexpand",
     "pyyaml",
     "pybind11",
-    "paddlelite",
     "paddleslim==2.3.4",
 ]
